Search data dictionary.

This function is used to search the IPEDS data dictionary.

ipeds_dict(
  search_string,
  search_col = c("all", "description", "varname", "filename"),
  exact_match = FALSE,
  limit = 10,
  confirm = FALSE,
  return_dict = FALSE,
  print_off = FALSE
)

Arguments

search_string

Character string for search. Can use regular expression for search. Must escape special characters, . \ | ( ) [ { ^ $ * + ?, with a doublebackslash \\.

search_col

Column to search. The default is to search all columns. Other options include: "varname", "description", "filename".

exact_match

Set to TRUE if you only want exact search string matches. Note that setting exact_match to TRUE may not work as expected if search string includes regular expressions.

limit

Only the first 10 dictionary items are returned by default. Increase to return more values. Set to Inf to return all items matched in search'

confirm

Use to confirm status of variable name in dictionary. Returns TRUE or FALSE.

return_dict

Return a tibble of the subset data dictionary.

print_off

Do not print to console; useful if you only want to return a tibble of dictionary values.

Interactive dictionary

When using ipeds_dict() interactively (the default), formatted output will be returned to the console.

> ipeds_dict("efydesom")

=====================================================================
VARIABLE: efydesom
=====================================================================

:::::::::::::::::::          DESCRIPTION          :::::::::::::::::::

Students enrolled in some but not all distance education courses

:::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::

../FILES

 |__ EFFY2020_DIST*
 |__ EFFY2021_DIST*
 |__ EFFY2022_DIST*
 |__ EFFY2023_DIST*

 * Denotes a long file in which institutions may have more than one
 record (UNITID values repeated across multiple rows).

=====================================================================
Printed information for 1 of out 1 variables.

Output includes:

Variable name

Each unique description found across dictionary files provided by NCES

All complete data files in which the variable with that description is found

A flag (*) if necessary to denote the data file is a long file

A note to indicate if more results are available than the limit are available

Examples

if (FALSE) { # \dontrun{
## simple search for "state" in any part of the dictionary
ipeds_dict("state")

## variable names starting with "st"
ipeds_dict("^st", search_col = "varname")

## return full dictionary (only recommended if not printing and
## storing in object)
df <- ipeds_dict(".", limit = Inf, print_off = TRUE, return_dict = TRUE)
} # }