Return a table of all currently available complete data files

Return a table with currently available IPEDS complete data files (see https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx). Unless called by the user on its own, the first ipeds_init() call in a session will call ipeds_file_table() to scrape the table from NCES website and store the results in a temporary directory. Subsequent data requests during the session will return the table in memory unless the user calls requests to redownload the table.

Usage

ipeds_file_table(redownload = FALSE, return_table = TRUE)

Arguments

redownload

Re-scrape NCES website to generate table of files. Defaults to FALSE.

return_table

Return table (Default = TRUE).

Value

A data frame with the following columns:

year: Survey year assigned to the file by NCES

survey: Survey file category

title: Survey file title

file: Survey file name

Examples

if (FALSE) { # \dontrun{
# First call in a session will scrape NCES website and return
ipeds_file_table()

# Second call in the session will return table in memory
ipeds_file_table()

# Will re-scrape NCES website
ipeds_file_table(redownload = TRUE)
} # }