Filter IPEDS data by variable values.

Filter downloaded IPEDS data. Simple filters separated by commas will be treated as if separated by an & (AND) boolean. More complex multi-part filters, such as those using an | (OR) boolean, must be input as a string or via an object which stores the string filter.

Usage

ipeds_filter(ipedscall, ...)

Arguments

ipedscall

Current list of parameters carried forward from prior functions in the chain (ignore)

...

Expressions to evaluate

Details

Non-standard evaluation may be used for simple filters and when ipeds_init(use_nse = TRUE). In all other cases, a string filter must be used.

Examples

if (FALSE) { # \dontrun{
# using non-standard evaluation, you can separate filters by a comma, which
# will serve as an (AND) boolean
ipeds_filter(stabbr == "KY", control < 3)
ipeds_filter(stabbr %in% c("KY","TN"))
ipeds_filter(grepl("community", instnm))

# more complex filters, such as those using an | (OR) operator, should be
# input as a string, either directly or as stored in an object
ipeds_filter("stabbr == 'KY' | stabbr == 'TN'")

filter_str <- "stabbr == 'KY' | stabbr == 'TN'"
ipeds_filter(filter_str)
} # }