Initialize chained request.

Initialize the data request. It should always be the first in the series of piped functions.

ipeds_init(local_dir = NA, use_revised_files = TRUE, use_nse = TRUE)

Arguments

local_dir

Local directory to check for complete IPEDS zip files that have already been downloaded. Newly downloaded files will be placed here as well. If unset, all files will be stored in tempdir().

use_revised_files

When IPEDS zip files contain revised data, use the revised files. Default is TRUE.

use_nse

Use non-standard evaluation in chain. Default is TRUE.

Examples

if (FALSE) { # \dontrun{
ipeds_init()

# set local directory to search first for existing IPEDS data files and in
# which to save any that must be downloaded
ipeds_init(local_dir = ".")

# use original files even if revise files exist
ipeds_init(use_revised_files = FALSE)

# use standard evaluation throughout the ipeds_*() chain
ipeds_init(use_nse = FALSE)
} # }